Refactor `compute_tid` to backend
