+++ onix/tests/test_utilities.py
@@ -124,71 +124,6 @@ class TestComputeSid(object):
         assert unsanitized_sid != sanitized_sid
 
 
-class TestComputeTid(object):
-
-    def setup_method(self, method):
-
-        try:
-            pokedex = json.load(open('.psdata/pokedex.json'))
-        except IOError:
-            pokedex = scrapers.scrape_battle_pokedex()
-
-        try:
-            aliases = json.load(open('.psdata/aliases.json'))
-        except IOError:
-            aliases = scrapers.scrape_battle_aliases()
-
-        self.sanitizer = utilities.Sanitizer(pokedex, aliases)
-
-        self.set_1 = Moveset([Forme('Blastoise-Mega', 'Mega Launcher',
-                                    PokeStats(361, 189, 276, 405, 268, 192)),
-                              Forme('Blastoise', 'Rain Dish',
-                                    PokeStats(361, 153, 236, 295, 248, 192))],
-                             'F', 'Blastoisinite',
-                             ['Water Spout', 'Aura Sphere', 'Dragon Pulse',
-                              'Dark Pulse'], 100, 255)
-        self.set_2 = Moveset([Forme('gardevoir', 'synchronize',
-                                    PokeStats(340, 121, 251, 286, 266, 197))],
-                             'u', 'choicescarf', ['healingwish'], 100, 255)
-
-        self.original_tid = utilities.compute_tid([self.set_1, self.set_2],
-                                                  self.sanitizer)
-
-    def test_tid_equivalence(self):
-
-        set_1a = Moveset([Forme('Blastoise-Mega', 'Mega Launcher',
-                                    PokeStats(360, 189, 276, 405, 268, 193)),
-                              Forme('Blastoise', 'Rain Dish',
-                                    PokeStats(360, 153, 236, 295, 248, 193))],
-                             'F', 'Blastoisinite',
-                             ['Water Spout', 'Aura Sphere', 'Dragon Pulse',
-                              'Dark Pulse'], 100, 255)
-
-        equivalent_tid = utilities.compute_tid([self.set_2, self.set_1],
-                                               self.sanitizer)
-        non_equivalent_tid = utilities.compute_tid([set_1a, self.set_2],
-                                                   self.sanitizer)
-
-        assert self.original_tid == equivalent_tid
-        assert self.original_tid != non_equivalent_tid
-
-    def test_compute_by_sids(self):
-
-        sid_1 = utilities.compute_sid(self.set_1, self.sanitizer)
-        sid_2 = utilities.compute_sid(self.set_2, self.sanitizer)
-
-        tid_by_sid = utilities.compute_tid([sid_1, sid_2])
-        reversed = utilities.compute_tid([sid_2, sid_1])
-
-        assert self.original_tid == tid_by_sid
-        assert self.original_tid == reversed
-
-    def test_invalid_type(self):
-
-        with pytest.raises(TypeError):
-            utilities.compute_tid([3, 6, 2])
-
-
 class TestDictToStats(object):
 
     def setup_method(self, method):
