+++ onix/utilities.py
@@ -4,9 +4,9 @@ from __future__ import print_function, division
 import collections
 import hashlib
 import json
-import pkg_resources
 import re
 
+import pkg_resources
 from past.builtins import basestring
 
 from onix.dto import Moveset, Forme, PokeStats
@@ -151,37 +151,6 @@ def compute_sid(moveset, sanitizer=None):
     # moveset_hash = moveset_hash[:16]
 
     return moveset_hash
-
-
-def compute_tid(team, sanitizer=None):
-    """
-    Computes the Team ID for the given group of movesets
-
-    Args:
-        team (:obj:`iterable` of :obj:`Moveset` or :obj:`str`) :
-            the team for which to compute the TID, represented either by their
-            movesets or by their SIDs
-        sanitizer (:obj:`Sanitizer`, optional):
-            if no sanitizer is provided, movesets are assumed to be already
-            sanitized. Otherwise, the provided ``Sanitizer`` is used to sanitize
-            the movesets.
-
-    Returns:
-        str: the corresponding Team ID
-    """
-    if isinstance(team[0], Moveset):
-        sids = [compute_sid(moveset, sanitizer) for moveset in team]
-    elif isinstance(team[0], str):
-        sids = team
-    else:
-        raise TypeError('team is neither an iterable of movesets nor SIDs')
-    sids = sorted(sids)
-    team_hash = hashlib.sha512(repr(sids).encode('utf-8')).hexdigest()
-
-    # may eventually want to truncate hash, e.g.
-    # team_hash = team_hash[:16]
-
-    return team_hash
 
 
 def stats_dict_to_dto(stats_dict):
